pycram-industrial demo: Launch file with prerequisites
Add launch file for ROS prerequisites of pycram-industrial demo. While
demos/pycram_industrial_demo/demo.py is not a ROS node and can't be
launched with ROS (it's Python3), it assumes that ROS is running
somewhere and that e.g. the rosbridge server is running. Add that to
the launch file.