Reasoning function to get the pose of a feature frame
Added the concept of "feature frames", frames local to an object which can perform context-dependent functions. A grasp point on an object is a feature frame, or the tip of a tool, or the handle of a cup. Feature frames are given as additional links in the URDF with fixed transforms from the object origin.